# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Class:
    properties:
      capacity:
        type: integer
      course:
        $ref: '#/definitions/models.Course'
      course_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        description: 'Contoh: "Basic English A"'
        type: string
      schedule:
        description: 'Contoh: "Senin & Rabu, 19:00 - 21:00"'
        type: string
      status:
        description: '''open'', ''full'', ''closed'''
        type: string
      updated_at:
        type: string
    type: object
  models.ClassPlacement:
    properties:
      class:
        $ref: '#/definitions/models.Class'
      class_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      placement_date:
        type: string
      registration:
        $ref: '#/definitions/models.Registration'
      registration_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Course:
    properties:
      classes:
        items:
          $ref: '#/definitions/models.Class'
        type: array
      created_at:
        type: string
      description:
        type: string
      duration:
        description: 'Contoh: "3 Bulan", "48 Jam"'
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      status:
        description: '''active'', ''inactive'''
        type: string
      updated_at:
        type: string
    type: object
  models.Payment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      payment_date:
        type: string
      payment_method:
        description: '''bank_transfer'', ''ewallet'', ''credit_card'''
        type: string
      registration:
        $ref: '#/definitions/models.Registration'
      registration_id:
        type: integer
      status:
        description: '''pending'', ''paid'', ''failed'', ''refunded'''
        type: string
      updated_at:
        type: string
    type: object
  models.Registration:
    properties:
      class_placement:
        $ref: '#/definitions/models.ClassPlacement'
      course:
        $ref: '#/definitions/models.Course'
      course_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      payment:
        $ref: '#/definitions/models.Payment'
      registration_date:
        type: string
      status:
        description: '''pending'', ''registered'', ''cancelled'', ''completed'''
        type: string
      student:
        $ref: '#/definitions/models.Student'
      student_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Student:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  services.CreateCourseRequest:
    properties:
      description:
        type: string
      duration:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      status:
        description: 'default: ''active'''
        type: string
    required:
    - duration
    - name
    - price
    type: object
...
  services.ProcessPaymentRequest:
    properties:
      amount:
        type: number
      payment_method:
        description: e.g. "bank_transfer", "ewallet", "credit_card"
        type: string
    required:
    - amount
    - payment_method
    type: object
...
  services.UpdateCourseRequest:
    properties:
      description:
        type: string
      duration:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      status:
        type: string
    required:
    - duration
    - name
    - price
    - status
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@englishcourse.com
    name: Novendra
  description: RESTful API backend untuk pendaftaran kursus bahasa Inggris dengan
    clean layered architecture.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: English Course Registration API
  version: "1.0"
paths:
...
  /courses:
    get:
      description: Mengembalikan katalog semua kursus bahasa Inggris yang tersedia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Course'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil daftar seluruh course
      tags:
      - Courses
    post:
      consumes:
      - application/json
      description: Menambahkan paket kursus baru beserta harga dan durasi
      parameters:
      - description: Payload data course
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Course'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Membuat katalog course baru
      tags:
      - Courses
  /courses/{id}:
    delete:
      description: Menghapus course berdasarkan ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Menghapus data course
      tags:
      - Courses
    get:
      description: Mengembalikan detail spesifik course beserta kelas yang dibuka
        di bawahnya
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Course'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil detail course beserta daftar kelasnya
      tags:
      - Courses
    put:
      consumes:
      - application/json
      description: Memperbarui nama, deskripsi, harga, durasi, atau status course
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload update course
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Course'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengubah data course
      tags:
      - Courses
...
  /payments:
    get:
      description: Mengembalikan daftar semua data transaksi payment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Payment'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil seluruh tagihan pembayaran
      tags:
      - Payments
  /payments/{id}:
    get:
      description: Mengembalikan informasi tagihan payment berdasarkan ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Payment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil detail tagihan pembayaran
      tags:
      - Payments
  /payments/{id}/pay:
    post:
      consumes:
      - application/json
      description: Melakukan pembayaran tagihan dan secara atomik mengubah status
        registrasi menjadi 'registered'
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload pembayaran
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.ProcessPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Payment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Memproses simulasi pembayaran tagihan
      tags:
      - Payments
...
swagger: "2.0"